pyproject.toml
[tool.poetry]
name = "homematch"
version = "0.1.0"
description = "Personalized Real Estate Agent using LangChain and Vector Databases"
authors = ["Mike Myat Min Han <mmhan@indeed.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# Core LangChain dependencies
langchain = "^0.3.10"
langchain-openai = "^0.3.10"
langchain-chroma = "^0.1.4"

# Vector Database
chromadb = "^0.5.1"

# Environment variable loading
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88